ES OFICIAL EMPRENDEDOR, ADMIN GUNCIONALIDAD Y FRONT HECHO. EDUQUE.

+++ src/main/java/com/usta/proyectointegrador/models/services/StartupServices.java
@@ -18,6 +18,6 @@ public interface StartupServices {
 
     List<StartupEntity> findByConvocatoria(ConvocatoriaEntity convocatoria);
 
-    List<StartupEntity> findByUsuario(Integer id);
+    List<StartupEntity> findByUsuario(Long id);
 
 }
